ADL: Refactor graphics code
